fix: move forms dir and fix login form
