+++ src/stories/Forms/RegisterForm/styled.tsx
